Added: Plugins!
There will be more documentation on this at a later date. Basically, to create a plugin, create a class that extends IRCPlugin (do not implement Plugin; only extend IRCPlugin or its subclasses). It has an onEnable and onDisable, along with some other useful methods for plugin authors. The plugin should be packaged into a jar with a plugin.yml, containing a main and a name String (see below). This jar should be put in the plugins folder (which must be created if it doesn’t exist) in the same directory as the bot’s jar.

Example (basic) plugin.yml:
```YAML
name: ExamplePlugin
main: com.example.MainPluginClass
```